requirements.txt
absl-py==1.4.0
astunparse==1.6.3
beautifulsoup4==4.11.1
CacheControl==0.12.11
cached-property==1.5.2
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.0
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
Cython==0.29.33
deepface==0.0.75
dill==0.3.6
dlib==19.24.0
dm-tree==0.1.8
etils==1.0.0
filelock==3.9.0
fire==0.5.0
firebase-admin==6.0.1
Flask==2.2.2
Flask-Ext==0.1
flatbuffers==23.1.4
fonttools==4.38.0
gast==0.4.0
gdown==4.6.0
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.9.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h5py==3.7.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
msgpack==1.0.4
mtcnn==0.1.1
mysql-connector-python==8.0.32
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
promise==2.3
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
termcolor==2.2.0
toml==0.10.2
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
